% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reedsolomon.R
\name{reedsolomon}
\alias{reedsolomon}
\title{Reed-Solomon Pixel-Level Algorithm for Remote Sensing Data Fusion}
\usage{
reedsolomon(x, a, stretch = FALSE, ...)
}
\arguments{
\item{x}{SpatRaster}

\item{stretch}{TRUE/FALSE Should all layers be stretch to the same range of values? It calls \code{\link[terra:stretch]{terra::stretch()}}}

\item{...}{Other parameters of  \code{\link[terra:stretch]{terra::stretch()}} and \code{\link[terra:app]{terra::app()}}}
}
\value{
SpatRaster
}
\description{
Reed-Solomon Pixel-Level Algorithm for Remote Sensing Data Fusion
}
\examples{
library(terra)
s1 = rast(system.file("ex/logo.tif", package = "terra"))
plot(s1)

r1 = reedsolomon(s1)
plot(r1)

s2 = s1 * 1.21
r2 = reedsolomon(s2)
s3 = s1 - 100
r3 = reedsolomon(s3)
}
\references{
Werner, P. A. (2020). Application of the Reed-Solomon Algorithm as a Remote Sensing Data Fusion Tool for Land Use Studies. Algorithms, 13(8), 188. MDPI AG. Retrieved from http://dx.doi.org/10.3390/a13080188
}
