% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reedsolomon_reverse_fusion.R
\name{reedsolomon_reverse_fusion}
\alias{reedsolomon_reverse_fusion}
\title{Reversal of Reed-Solomon Pixel-Level Algorithm for Remote Sensing Data Fusion}
\usage{
reedsolomon_reverse_fusion(x, a, k, ...)
}
\arguments{
\item{x}{SpatRaster, the output of \code{\link[=reedsolomon_fusion]{reedsolomon_fusion()}}}

\item{a}{Maximum theoretical number of categories (integer values)}

\item{k}{Expected number of raster layers}

\item{...}{Other parameters of \code{\link[terra:app]{terra::app()}}}
}
\value{
SpatRaster
}
\description{
Reversal of Reed-Solomon Pixel-Level Algorithm for Remote Sensing Data Fusion
}
\examples{
library(terra)
s1 = rast(system.file("ex/logo.tif", package = "terra"))
plot(s1)

r1 = reedsolomon_fusion(s1, a = 256)
plot(r1)

rev_r1 = reedsolomon_reverse_fusion(r1, a = 256, k = 3)
all.equal(values(s1), values(rev_r1))
}
